fix(Chip): correct dependencies, move stories
