--- libs/ui/chip/.storybook/preview-head.html
@@ -1,13 +1,0 @@
-<link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,400;0,500;0,700;1,400&display=swap">
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@terminus/ui-styles/terminus-ui.min.css">
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@terminus/design-tokens/css/design-tokens.min.css">
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@terminus/design-tokens/css/library-design-tokens.min.css">
-
-<style type="text/css">
-  html {
-    font-family: var(--ts-typography-compound-body);
-  }
-  .sb-show-main {
-    box-sizing: border-box;
-  }
-</style>
