--- libs/ui/chip/.storybook/webpack.config.js
@@ -1,6 +1,0 @@
-const rootWebpackConfig = require('../../../../.storybook/webpack.config');
-// Export a function. Accept the base config as the only param.
-module.exports = async ({ config, mode }) => {
-  config = await rootWebpackConfig({ config, mode });
-  return config;
-};
