+++ libs/ui/chip/schematics/ng-add/index.ts
@@ -20,13 +20,17 @@ export const ngAdd = () => (tree: Tree, context: SchematicContext): Tree => {
     '@angular/forms: ^9.1.0',
     '@angular/material: ^9.2.4',
     '@angular/platform-browser: ^9.1.0',
+    '@fortawesome/angular-fontawesome: ^0.6.0',
+    '@fortawesome/fontawesome-svg-core: ^1.2.30',
+    '@fortawesome/pro-solid-svg-icons: ^5.14.0',
     '@terminus/design-tokens: ^3.1.0',
     '@terminus/fe-utilities: ^1.0.0',
     '@terminus/ui-chip: ^2.0.4',
-    '@terminus/ui-icon: ^2.0.5',
+    '@terminus/ui-icon: ^2.1.5',
+    '@terminus/ui-spacing: ^1.1.3',
     '@terminus/ui-styles: ^1.0.13',
-    'date-fns: 2.14.0',
     '@terminus/ui-utilities: ^1.0.9',
+    'date-fns: 2.14.0',
     'rxjs: ^6.5.0',
     'tslib: ^1.10.0',
   ].map(p => {
