+++ stories/ui-chip/.storybook/main.js
@@ -1,9 +1,9 @@
-const addonConfigs = require('../../../../.storybook/utilities/addon-configs');
+const addonConfigs = require('../../../.storybook/utilities/addon-configs');
 
 module.exports = {
   addons: [
     ...addonConfigs.addonsNoEssentials,
     ...addonConfigs.essentialsNoDocs,
   ],
-  stories: ['../src/**/*.stories.@(js|ts|mdx)'],
+  stories: ['../**/*.stories.@(js|ts|mdx)'],
 };
