+++ stories/ui-chip/.storybook/tsconfig.json
@@ -1,8 +1,8 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "../../../libs/ui/chip/tsconfig.json",
   "compilerOptions": {
     "emitDecoratorMetadata": true
   },
-  "exclude": ["../src/test.ts", "../**/*.spec.ts"],
-  "include": ["../src/**/*"]
+  "exclude": ["../../../libs/ui/chip/src/test.ts", "../../../libs/ui/chip/**/*.spec.ts"],
+  "include": ["../../../libs/ui/chip/src/**/*"]
 }
